delete authorization

+++ app/Http/Controllers/PuppyController.php
@@ -104,7 +104,14 @@ class PuppyController extends Controller
     // ------------------------------
     public function destroy(Request $request, Puppy $puppy)
     {
+        sleep(2);
+
         $imagePath = str_replace('/storage/', '', $puppy->image_url);
+
+        if ($request->user()->cannot('delete', $puppy)) {
+            return back()
+                ->withErrors(['error' => 'You do not have permission to delete this puppy.']);
+        }
 
         $puppy->delete();
